+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Bump MDC to v1.3.0
 - Change `PatientSearchResultsScreen` to a fragment
 - Bump ConstraintLayout to v2.0.4
+- Add `Widget.Simple.TextField.Layout.PatientEntry.AutoComplete` style
 - [In Progress: 20 Jan 2021] Material Theming Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding