+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Change `PatientSearchResultsScreen` to a fragment
 - Bump ConstraintLayout to v2.0.4
 - Add `Widget.Simple.TextField.Layout.PatientEntry.AutoComplete` style
+- Bump Kotlin to v1.4.30
 - [In Progress: 20 Jan 2021] Material Theming Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding